LSP: Move lspmanager to preferences module
Most of its code belongs there, although not all.
