+++ spyder/plugins/editor/widgets/tests/test_warnings.py
@@ -17,7 +17,7 @@ from spyder.utils.qthelpers import qapplication
 from spyder.plugins.editor.widgets.codeeditor import CodeEditor
 from spyder.py3compat import to_binary_string
 from spyder.utils.codeanalysis import check_with_pyflakes, check_with_pep8
-from spyder.plugins.lspmanager import LSPManager
+from spyder.preferences.languageserver import LSPManager
 from spyder.plugins.editor.lsp import LSPEventTypes
 from spyder.py3compat import PY2
